Terminate all middleware.
On Laravel, the application middleware was being terminated correctly, however, route
middleware was not. Laravel route middleware is not terminated correctly.

On Lumen, route middleware does not terminate, so it was not added. However, because
we stripped the application middleware from Lumen we needed to manually terminate
the application middleware.

Reference: #721