Add any new parent requests prior to queuing up a new internal request.
This mainly applies to testing. When crawling pages and several requests are
made, only the initial request is added to the dispatchers request stack.
This means that if two or more requests are made prior to dispatching
an internal request, the parent request in the request stack is
incorrect and no longer refers to the actual parent of the
internal request.

The solution here was to add the immediate parent request if the intitial
request in the stack did not match the current application request.

Reference issue #808.