Revert 9b04bfa, instead decode the JSON into a new Response.
This change reverts the commit so that a JsonResponse is allowed, although
it is still not prefered. The JsonResponse is converted into a normal
Response prior to being morphed, this results in the content being
decoded into a plain array. This provides consistency across
the board.

So if you internally request an endpoint that returns a JsonResponse
of a model or collection, you'll get an array, not the actual model
or collection. Just worth pointing out.

Reference issue #879.