When running in console watch for request rebinding and set the route resolver.
When Lumen is running in the console there is no external request which means the service
bindings for the request (Application::registerRequestBindings) are never run so the
route resolver is never set. This is different from Laravel in that the route
resolver is set once the route has been matched from the routing dispatcher.
To get around this we'll watch for the rebinding of rebinding of the
request object and simply use reflection to grab the currentRoute
property and return it for the route resolver.

Reference issue #881.