Further work on refining the Laravel adapter to use its own router instance.
Because the Laravel adapter calls the Router::setRoutes method the routes container
binding is set again causing the rebinding to trigger and update the routes on the
applications UrlGenerator. To resolve this we will manually set the routes on the
container to trigger the binding and reset the UrlGenerators routes.

Reference issue #918.