Normalize the FastRoute static routes between multiple versions.
There were some changes in FastRoute that resulting in static routes array
being slightly different (the method/uri were swapped in the keys). This
resulted in some odd results between Lumen 5.1 and 5.2.

This fix now normalizes the static routes from 5.1 to make the array
the same as 5.2. The routes are then added properly when fetching
the iterable routes.

Reference issue #868 and #945.