Allow internal requests via commands
Currently internal requests can't be made from within a command. When `$this->getRootRequest()->root()` runs it returns `http://:` because it is not able to resolve a host from `$_SERVER` (have a look at `Symfony\Component\HttpFoundation\Request@getHost`). This will prevent that from getting prefixed and allow requests from the command line.